requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
altair==5.4.1
antlr4-python3-runtime==4.9.3
async-timeout==5.0.1
attrs==25.3.0
blinker==1.8.2
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
diffusers==0.33.1
einops==0.8.1
filelock==3.16.1
frozenlist==1.5.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
hf-xet==1.1.2
huggingface-hub==0.32.3
Hydra==2.5
hydra-core==1.3.2
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
iopath==0.1.10
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
latent-diffusion @ file:///home/llt/repos/idles/runway-stable-diffusion-inpainting
lightning-utilities==0.11.9
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
narwhals==1.41.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.41
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opencv-python==4.11.0.86
packaging==24.2
pandas==2.0.3
pillow==10.4.0
pkgutil_resolve_name==1.3.10
portalocker==3.0.0
propcache==0.2.0
protobuf==5.29.5
pyarrow==17.0.0
pydeck==0.9.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.20.1
safetensors==0.5.3
scipy==1.10.1
six==1.17.0
smmap==5.0.2
streamlit==1.40.1
streamlit-drawable-canvas==0.9.3
sympy==1.13.3
tenacity==9.0.0
tokenizers==0.20.3
toml==0.10.2
torch==2.4.1
torch-fidelity==0.3.0
torchmetrics==1.5.2
torchvision==0.19.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.46.3
triton==3.0.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
watchdog==4.0.2
yarl==1.15.2
zipp==3.20.2
